requirements.txt
numpy>=1.23
sounddevice>=0.4.5
soundfile>=0.11.0
pyttsx3>=2.90
pydub>=0.25.1
python-dotenv>=1.0.0
requests>=2.28.0
openai-whisper>=20230314

# Notes:
# - FFmpeg is required by whisper/pydub for some audio operations; install FFmpeg separately and add to PATH.
# - If you use a different speech library, adjust this file accordingly.